backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
openpyxl>=3.1.0
